cleanup: generalized `Roles` component
